Consolidate duplicated implementation of service specification
Previously we had two implementations of the service specification:
one in interfaces.base.py and the other one in
model.service_specification.py.

This removes the implementation in interfaces.base.py and adapt the
correponding usages to the one in model.service_specification.py.

Fixes #523
