+++ server/app/interfaces/discovery.py
@@ -5,7 +5,7 @@ This module implements the Discovery interface defined in the
 """
 
 import json
-from typing import Dict, List, Set
+from typing import Dict, List, Set, Type
 
 import werkzeug.exceptions
 from basyx.aas import model
@@ -14,8 +14,14 @@ from werkzeug.wrappers import Request, Response
 
 from app import model as server_model
 from app.adapter import jsonization
-from app.interfaces.base import BaseWSGIApp, HTTPApiDecoder
+from app.interfaces.base import BaseWSGIApp, HTTPApiDecoder, APIResponse
 from app.util.converters import IdentifierToBase64URLConverter, base64url_decode
+from app.model import ServiceSpecificationProfileEnum, ServiceDescription
+
+SUPPORTED_PROFILES: ServiceDescription = ServiceDescription([
+    ServiceSpecificationProfileEnum.DISCOVERY_FULL,
+    ServiceSpecificationProfileEnum.DISCOVERY_READ,
+])
 
 
 class DiscoveryStore:
@@ -90,6 +96,7 @@ class DiscoveryAPI(BaseWSGIApp):
                 Submount(
                     base_path,
                     [
+                        Rule("/description", methods=["GET"], endpoint=self.get_description),
                         Rule(
                             "/lookup/shellsByAssetLink",
                             methods=["POST"],
@@ -120,8 +127,11 @@ class DiscoveryAPI(BaseWSGIApp):
             strict_slashes=False,
         )
 
+    def get_description(self, request: Request, url_args: Dict, response_t: Type[APIResponse], **_kwargs) -> Response:
+        return response_t(SUPPORTED_PROFILES.to_dict())
+
     def get_all_aas_ids_by_asset_link(
-            self, request: Request, url_args: dict, response_t: type, **_kwargs
+            self, request: Request, url_args: dict, response_t: Type[APIResponse], **_kwargs
     ) -> Response:
         asset_ids_param = request.args.get("assetIds", "")
         if not asset_ids_param:
@@ -152,7 +162,7 @@ class DiscoveryAPI(BaseWSGIApp):
         return response_t(list(paginated_slice), cursor=cursor)
 
     def search_all_aas_ids_by_asset_link(
-        self, request: Request, url_args: dict, response_t: type, **_kwargs
+        self, request: Request, url_args: dict, response_t: Type[APIResponse], **_kwargs
     ) -> Response:
         asset_links = HTTPApiDecoder.request_body_list(request, server_model.AssetLink, False)
         matching_aas_keys = set()
@@ -163,13 +173,15 @@ class DiscoveryAPI(BaseWSGIApp):
         return response_t(list(paginated_slice), cursor=cursor)
 
     def get_all_specific_asset_ids_by_aas_id(
-        self, request: Request, url_args: dict, response_t: type, **_kwargs
+        self, request: Request, url_args: dict, response_t: Type[APIResponse], **_kwargs
     ) -> Response:
         aas_identifier = str(url_args["aas_id"])
         asset_ids = self.persistent_store.get_all_specific_asset_ids_by_aas_id(aas_identifier)
         return response_t(asset_ids)
 
-    def post_all_asset_links_by_id(self, request: Request, url_args: dict, response_t: type, **_kwargs) -> Response:
+    def post_all_asset_links_by_id(
+        self, request: Request, url_args: dict, response_t: Type[APIResponse], **_kwargs
+    ) -> Response:
         aas_identifier = str(url_args["aas_id"])
         specific_asset_ids = HTTPApiDecoder.request_body_list(request, model.SpecificAssetId, False)
         self.persistent_store.add_specific_asset_ids_to_aas(aas_identifier, specific_asset_ids)
@@ -178,7 +190,9 @@ class DiscoveryAPI(BaseWSGIApp):
         updated = {aas_identifier: self.persistent_store.get_all_specific_asset_ids_by_aas_id(aas_identifier)}
         return response_t(updated)
 
-    def delete_all_asset_links_by_id(self, request: Request, url_args: dict, response_t: type, **_kwargs) -> Response:
+    def delete_all_asset_links_by_id(
+        self, request: Request, url_args: dict, response_t: Type[APIResponse], **_kwargs
+    ) -> Response:
         aas_identifier = str(url_args["aas_id"])
         self.persistent_store.delete_specific_asset_ids_by_aas_id(aas_identifier)
         for key in list(self.persistent_store.asset_id_to_aas_ids.keys()):
